breakout more modules, added jshint
