+++ .jshintignore
@@ -1,0 +1,3 @@
+stubs
+dist
+node_modules
