+++ debug.log
@@ -1,0 +1,2 @@
+[0221/232558:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:6004
+[0221/232558:ERROR:node_debugger.cc(86)] Cannot start debugger server
