+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "target": "ES6"
+    },
+    "exclude": [
+        "node_modules", "dist"
+    ]
+}
