+++ src/PFDB.js
@@ -1,0 +1,74 @@
+export default {
+	/* Pathfinder RULE constants for parsing */
+	unarmedAttacksRegExp : /unarmed|strike|punch|palm|flurry|blow|touch/i,
+	combatManeuversRegExp : /bull\s*rush|trip|disarm|dirty\s*trick|drag|grapple|overrun|reposition|steal|sunder|trip/i,
+	primaryNaturalAttacksRegExp : /bite|claw|gore|slam|sting|talon/i,
+	importantFeatRegExp : /Weapon Fine|^Run$|Enduran|Defensive Combat Train|Agile Maneuv|Arcane Arm|Combat Cast|Critical Focus|Skill Focus|Critical|Intimidating Prow/i,
+	spAttackAttacks : /blooddrain|touch|energydrain|bleed|burn|constrict|trample|engulf|heat|powerfulcharge|swallowwhole/i,
+	spAttackAttacksPreProcess : /rake|rend|web/i,
+	abilitySrch : /str|dex|con|int|wis|cha/i,
+	cmbMonsterSrch : /swallowwhole|tongue|pull|drag|grab|push/i,
+	casterDefaultAbility : {
+		sorcerer: 'cha',
+		wizard: 'int',
+		cleric: 'wis',
+		bard: 'cha',
+		druid: 'wis',
+		paladin: 'cha',
+		ranger: 'wis',
+		oracle: 'cha',
+		witch: 'int',
+		alchemist: 'int',
+		summoner: 'cha',
+		inquisitor: 'wis',
+		magus: 'int'
+	},
+	creatureTypeClassSkills : {
+		aberration: ['Acrobatics', 'Climb', 'Escape-Artist', 'Fly', 'Intimidate', 'Knowledge', 'Perception', 'Spellcraft', 'Stealth', 'Survival', 'Swim'],
+		animal: ['Acrobatics', 'Climb', 'Fly', 'Perception', 'Stealth', 'Swim'],
+		dragon: ['Appraise', 'Bluff', 'Climb', 'Craft', 'Diplomacy', 'Fly', 'Heal', 'Intimidate', 'Linguistics', 'Knowledge', 'Perception', 'Sense-Motive', 'Spellcraft', 'Stealth', 'Survival', 'Swim', 'Use-Magic-Device'],
+		fey: ['Appraise', 'Bluff', 'Climb', 'Craft', 'Diplomacy', 'Disguise', 'Escape-Artist', 'Fly', 'Knowledge-Geography', 'Knowledge-Local', 'Knowledge-Nature', 'Perception', 'Perform', 'Sense-Motive', 'Sleight-of-Hand', 'Stealth', 'Swim', 'Use-Magic-Device'],
+		monstrous: ['Climb', 'Craft', 'Fly', 'Intimidate', 'Perception', 'Ride', 'Stealth', 'Survival', 'Swim'],
+		humanoid: ['Climb', 'Craft', 'Handle-Animal', 'Heal', 'Profession', 'Ride', 'Survival'],
+		magicalbeast: ['Acrobatics', 'Climb', 'Fly', 'Perception', 'Steatlh', 'Swim'],
+		outsider: ['Bluff', 'Craft', 'Knowledge-Planes', 'Perception', 'Sense-Motive', 'Stealth'],
+		plant: ['Perception', 'Stealth'],
+		undead: ['Climb', 'Disguise', 'Fly', 'Intimidate', 'Knowledge-Arcana', 'Knowledge-Religion', 'Perception', 'Sense-Motive', 'Spellcraft', 'Stealth']
+	},
+	creatureSubtypeClassSkills :{
+		air: ['Fly'],
+		giant: ['Intimidate', 'Perception'],
+		goblinoid: ['Stealth'],
+		inevitable: ['Acrobatics', 'Diplomacy', 'Intimidate', 'Survival'],
+		water: ['Swim'],
+		robot: ['Climb', 'Disable-Device', 'Fly', 'Knowledge', 'Linguistics', 'Perception', 'Sense-Motive']
+	},
+	specialAttackDCAbilityBase : {
+		'breathweapon': 'CON',
+		'burn': 'CON',
+		'curse': 'CHA',
+		'disease': 'CON',
+		'distraction': 'CON',
+		'emotionaura': 'CHA',
+		'energydrain': 'CHA',
+		'entrap': 'CON',
+		'fear': 'CHA',
+		'fearaura': 'CHA',
+		'fearcone': 'CHA',
+		'fearray': 'CHA',
+		'frightfulpresence': 'CHA',
+		'gaze': 'CHA',
+		'mentalstaticaura': 'CHA',
+		'paralysis': 'CON',
+		'poison': 'CON',
+		'stench': 'CON',
+		'trample': 'STR',
+		'web': 'CON',
+		'whirlwind': 'STR'
+	},
+	monsterRulesPlusAttacks : ['grab', 'trip', 'engulf', 'swallow whole'],
+	naturalAttackRegExp : /arm|bite|claw|foreclaw|gore|hoof|kick|leg|pincer|quill|root|slam|slap|snake\s*bite|spike|spine|splinter|sting|tail|tail\s*slap|talon|tendril|tentacle|thorn|tongue|vine|wing|buffet/i,
+	skillBonusFeatRegExp : /Acrobatic|Uncanny Alertness|Alertness|Animal Affinity|Athletic|Breadth of Experience|Brewmaster|Deceitful|Deft Hands|Deny the Reaper|Forgotten Past|Innocent Blood|Monkey Moves|Monkey Style|Monument Builder|No Name|Persuasive|Scavenger.s Eye|Sea Legs|Self.Sufficient|Sharp Senses|Stealthy|Sure and Fleet|Voice of the Sibyl/i,
+	lessImportantCombatFeatRegExp : /Power Attack|Double Slice|Greater Weapon of the Chosen|Spell Focus|Toughness/i,
+	monsterRules : /spells|damage reduction/i
+}
