+++ src/index.js
@@ -1,10 +1,14 @@
 import _ from 'underscore';
-import on from '../stubs/on';
+import {on,randomInteger,getsectionIDs,getTranslationByKey,getAttrs,setAttrs,removeRepeatingRow,generateRowID} from '../stubs/on';
 import TAS from 'exports-loader?TAS!TheAaronSheet';
 import * as ExExp from './ExExp';
 import {PFLog} from './PFLog';
 import * as SWUtils from './SWUtils';
-
+import PFConst from './PFConst';
+import PFDB from './PFDB';
+import * as PFUtils from './PFUtils';
+import * as PFMacros from './PFMacros';
+//how to run this before anything? even before stuff is loaded?
 TAS.config({
  logging: {
    info: true,
@@ -13,1584 +17,7 @@ TAS.config({
 });
 TAS.debugMode();
 
-/* //UNCOMMENT FOR JS LINT
-console.log('%c•!!!!!!!!!!!!!!!!!!IF YOU SEE THIS YOU FORGOT TO UNCOMMENT THE TEST CODE FOR JS LINT!!!!!!!•', 'background: linear-gradient(to right,red,white,white,red); color:black;text-shadow: 0 0 8px white;');
-var randomInteger = function () {'use strict';};
-var getSectionIDs = function () {'use strict';};
-var getTranslationByKey = function () {'use strict';};
-var getAttrs = function () {'use strict';};
-var setAttrs = function () {'use strict';};
-var on = function () {'use strict';};
-var removeRepeatingRow = function () {'use strict';};
-var generateRowID = function () {'use strict';};
-var _ = _ || (function () {'use strict';
-return { dummy  : 0};
-}());
-*/
 
-var PFConst = {
-	/* Pathfinder SHEET constants */
-	version: 1.15,
-	/***************************************Lists of Fields ************************************************************/
-	//add any new repeating sections here. This is the word after "repeating_"
-	repeatingSections: ["weapon", "ability", "class-ability", "feat", "racial-trait", "trait", "item", "npc-spell-like-abilities", "mythic-ability", "mythic-feat", "buff", "spells"],
-	//repeating sections that have used and used|max and max-calculation fields
-	repeatingMaxUseSections: ["class-ability", "feat", "racial-trait", "trait", "mythic-ability", "mythic-feat", "ability"],
-
-	//attribute of a dropdown mapped to attribute to write evaluated number to.
-	//all simple dropdowns that do not need to call any other function when evaluating besides setDropdownValue and findAbilityInString
-	dropdowns: {
-		"HP-ability": "HP-ability-mod",
-		"init-ability": "init-ability-mod",
-		"Fort-ability": "Fort-ability-mod",
-		"Ref-ability": "Ref-ability-mod",
-		"Will-ability": "Will-ability-mod",
-		"melee-ability": "melee-ability-mod",
-		"melee2-ability": "melee2-ability-mod",
-		"ranged-ability": "ranged-ability-mod",
-		"ranged2-ability": "ranged2-ability-mod",
-		"CMB-ability": "CMB-ability-mod",
-		"CMB2-ability": "CMB2-ability-mod"
-	},
-	//attribute of a macro, mapped to attribute to write evaluation to
-	//all simple macros that do not need to call other functions besides evaluateAndSetNumber
-	equationMacros: {
-		"init-misc": "init-misc-mod",
-		"HP-formula-macro-text": "HP-formula-mod",
-		"Max-Skill-Ranks-Misc": "Max-Skill-Ranks-mod",
-		"SR-macro-text": "SR",
-		"spellclass-0-SP_misc": "spellclass-0-SP-mod",
-		"spellclass-1-SP_misc": "spellclass-1-SP-mod",
-		"spellclass-2-SP_misc": "spellclass-2-SP-mod"
-	},
-	//the 3 spell classes at top of spells page
-	spellClassIndexes: ["0", "1", "2"],
-	silentParams : {silent:true},
-	minusreg : /\-|\u2013|\u2014|\u2212|\u02d7/,
-	critreg : /(\d+)[\-|\u2013|\u2014|\u2212|\u02d7]20\/[x\u00d7](\d+)/,
-	diceDiereg : /(\d+)d(\d+)\s*([\+|\-|\u2013|\u2014|\u2212|\u02d7]{0,1})\s*(\d*)/
-};
-var PFDB = {
-	/* Pathfinder RULE constants for parsing */
-	unarmedAttacksRegExp : /unarmed|strike|punch|palm|flurry|blow|touch/i,
-	combatManeuversRegExp : /bull\s*rush|trip|disarm|dirty\s*trick|drag|grapple|overrun|reposition|steal|sunder|trip/i,
-	primaryNaturalAttacksRegExp : /bite|claw|gore|slam|sting|talon/i,
-	importantFeatRegExp : /Weapon Fine|^Run$|Enduran|Defensive Combat Train|Agile Maneuv|Arcane Arm|Combat Cast|Critical Focus|Skill Focus|Critical|Intimidating Prow/i,
-	spAttackAttacks : /blooddrain|touch|energydrain|bleed|burn|constrict|trample|engulf|heat|powerfulcharge|swallowwhole/i,
-	spAttackAttacksPreProcess : /rake|rend|web/i,
-	abilitySrch : /str|dex|con|int|wis|cha/i,
-	cmbMonsterSrch : /swallowwhole|tongue|pull|drag|grab|push/i,
-	casterDefaultAbility : {
-		sorcerer: 'cha',
-		wizard: 'int',
-		cleric: 'wis',
-		bard: 'cha',
-		druid: 'wis',
-		paladin: 'cha',
-		ranger: 'wis',
-		oracle: 'cha',
-		witch: 'int',
-		alchemist: 'int',
-		summoner: 'cha',
-		inquisitor: 'wis',
-		magus: 'int'
-	},
-	creatureTypeClassSkills : {
-		aberration: ['Acrobatics', 'Climb', 'Escape-Artist', 'Fly', 'Intimidate', 'Knowledge', 'Perception', 'Spellcraft', 'Stealth', 'Survival', 'Swim'],
-		animal: ['Acrobatics', 'Climb', 'Fly', 'Perception', 'Stealth', 'Swim'],
-		dragon: ['Appraise', 'Bluff', 'Climb', 'Craft', 'Diplomacy', 'Fly', 'Heal', 'Intimidate', 'Linguistics', 'Knowledge', 'Perception', 'Sense-Motive', 'Spellcraft', 'Stealth', 'Survival', 'Swim', 'Use-Magic-Device'],
-		fey: ['Appraise', 'Bluff', 'Climb', 'Craft', 'Diplomacy', 'Disguise', 'Escape-Artist', 'Fly', 'Knowledge-Geography', 'Knowledge-Local', 'Knowledge-Nature', 'Perception', 'Perform', 'Sense-Motive', 'Sleight-of-Hand', 'Stealth', 'Swim', 'Use-Magic-Device'],
-		monstrous: ['Climb', 'Craft', 'Fly', 'Intimidate', 'Perception', 'Ride', 'Stealth', 'Survival', 'Swim'],
-		humanoid: ['Climb', 'Craft', 'Handle-Animal', 'Heal', 'Profession', 'Ride', 'Survival'],
-		magicalbeast: ['Acrobatics', 'Climb', 'Fly', 'Perception', 'Steatlh', 'Swim'],
-		outsider: ['Bluff', 'Craft', 'Knowledge-Planes', 'Perception', 'Sense-Motive', 'Stealth'],
-		plant: ['Perception', 'Stealth'],
-		undead: ['Climb', 'Disguise', 'Fly', 'Intimidate', 'Knowledge-Arcana', 'Knowledge-Religion', 'Perception', 'Sense-Motive', 'Spellcraft', 'Stealth']
-	},
-	creatureSubtypeClassSkills :{
-		air: ['Fly'],
-		giant: ['Intimidate', 'Perception'],
-		goblinoid: ['Stealth'],
-		inevitable: ['Acrobatics', 'Diplomacy', 'Intimidate', 'Survival'],
-		water: ['Swim'],
-		robot: ['Climb', 'Disable-Device', 'Fly', 'Knowledge', 'Linguistics', 'Perception', 'Sense-Motive']
-	},
-	specialAttackDCAbilityBase : {
-		'breathweapon': 'CON',
-		'burn': 'CON',
-		'curse': 'CHA',
-		'disease': 'CON',
-		'distraction': 'CON',
-		'emotionaura': 'CHA',
-		'energydrain': 'CHA',
-		'entrap': 'CON',
-		'fear': 'CHA',
-		'fearaura': 'CHA',
-		'fearcone': 'CHA',
-		'fearray': 'CHA',
-		'frightfulpresence': 'CHA',
-		'gaze': 'CHA',
-		'mentalstaticaura': 'CHA',
-		'paralysis': 'CON',
-		'poison': 'CON',
-		'stench': 'CON',
-		'trample': 'STR',
-		'web': 'CON',
-		'whirlwind': 'STR'
-	},
-	monsterRulesPlusAttacks : ['grab', 'trip', 'engulf', 'swallow whole'],
-	naturalAttackRegExp : /arm|bite|claw|foreclaw|gore|hoof|kick|leg|pincer|quill|root|slam|slap|snake\s*bite|spike|spine|splinter|sting|tail|tail\s*slap|talon|tendril|tentacle|thorn|tongue|vine|wing|buffet/i,
-	skillBonusFeatRegExp : /Acrobatic|Uncanny Alertness|Alertness|Animal Affinity|Athletic|Breadth of Experience|Brewmaster|Deceitful|Deft Hands|Deny the Reaper|Forgotten Past|Innocent Blood|Monkey Moves|Monkey Style|Monument Builder|No Name|Persuasive|Scavenger.s Eye|Sea Legs|Self.Sufficient|Sharp Senses|Stealthy|Sure and Fleet|Voice of the Sibyl/i,
-	lessImportantCombatFeatRegExp : /Power Attack|Double Slice|Greater Weapon of the Chosen|Spell Focus|Toughness/i,
-	monsterRules : /spells|damage reduction/i
-};
-var PFUtils = PFUtils || (function () {
-	'use strict';
-	/****************************SYNCHRONOUS UTILITIES ***********************************
-	NO asynchronous FUNCTIONS SHOULD GO HERE
-	************************************************************************************** */
-	/** findAbilityInString - returns the attribute referenced by a dropdown option value.
-	* Looks at a string for instances of an ability modifier DEX-mod, STR-mod,  etc and returns the modifier it finds.
-	* if none are found, or if the first character is "0", return ""
-	* NOTE: YOU MUST PUT ANY NEW DROPDOWN VALUES HERE! 
-	* (if they are references to other fields. obviously, dropdowns with 0, 1, 2 as values are not needed here)
-	*@param {string} stringToSearch the value of the dropdown option selected
-	*@returns {string} the attribute referenced by a dropdown option value.
-	*/
-	var findAbilityInString = function (stringToSearch) {
-		if (!stringToSearch) {
-			return "";
-		}
-		if (stringToSearch.slice(0, 1) === "0") {
-			return "";
-		}
-		if (/str.mod/i.test(stringToSearch)) {
-			return "STR-mod";
-		}
-		if (/dex.mod/i.test(stringToSearch)) {
-			return "DEX-mod";
-		}
-		if (/con.mod/i.test(stringToSearch)) {
-			return "CON-mod";
-		}
-		if (/int.mod/i.test(stringToSearch)) {
-			return "INT-mod";
-		}
-		if (/wis.mod/i.test(stringToSearch)) {
-			return "WIS-mod";
-		}
-		if (/cha.mod/i.test(stringToSearch)) {
-			return "CHA-mod";
-		}
-		if (/melee2/i.test(stringToSearch)) {
-			return "attk-melee2";
-		}
-		if (/melee/i.test(stringToSearch)) {
-			return "attk-melee";
-		}
-		if (/ranged2/i.test(stringToSearch)) {
-			return "attk-ranged2";
-		}
-		if (/ranged/i.test(stringToSearch)) {
-			return "attk-ranged";
-		}
-		if (/cmb2/i.test(stringToSearch)) {
-			return "CMB2";
-		}
-		if (/cmb/i.test(stringToSearch)) {
-			return "CMB";
-		}
-		if (/str/i.test(stringToSearch)) {
-			return "STR";
-		}
-		if (/dex/i.test(stringToSearch)) {
-			return "DEX";
-		}
-		if (/con/i.test(stringToSearch)) {
-			return "CON";
-		}
-		if (/int/i.test(stringToSearch)) {
-			return "INT";
-		}
-		if (/wis/i.test(stringToSearch)) {
-			return "WIS";
-		}
-		if (/cha/i.test(stringToSearch)) {
-			return "CHA";
-		}
-		if (/npc.type/i.test(stringToSearch)) {
-			return "npc-type";
-		}
-		if (/race/i.test(stringToSearch)) {
-			return "race";
-		}
-		if (/class.0.level/i.test(stringToSearch)) {
-			return "class-0-level";
-		}
-		if (/\{level\}/i.test(stringToSearch)) {
-			return "level";
-		}
-		if (/npc.hd.num/i.test(stringToSearch)) {
-			return "npc-hd-num";
-		}
-		if (/class.1.level/i.test(stringToSearch)) {
-			return "class-1-level";
-		}
-		if (/class.2.level/i.test(stringToSearch)) {
-			return "class-2-level";
-		}
-		if (/class.3.level/i.test(stringToSearch)) {
-			return "class-3-level";
-		}
-		if (/class.4.level/i.test(stringToSearch)) {
-			return "class-4-level";
-		}
-		if (/class.5.level/i.test(stringToSearch)) {
-			return "class-5-level";
-		}
-	},
-	/** calculateSpellRanges - returns {close:x, medium:y , long:z} for casterlevel 
-	*@param {int} casterlevel level of caster
-	*@returns {jsobject} mapping like this: {close:int,medium:int,long:int}
-	*/
-	calculateSpellRanges = function (casterlevel) {
-		casterlevel = casterlevel || 0;
-		return {
-			close: 25 + (5 * (Math.floor(casterlevel / 2))),
-			medium: 100 + (10 * casterlevel),
-			"long": 400 + (40 * casterlevel)
-		};
-	},
-	/**findSpellRange -  calculates range number based on spell settings
-	* @param {int} customRangeVal value that is in the custom range field, for "per level" or "custom" choices
-	* @param {string} rangeDropdown selected value from spell range dropdown
-	* @param {int} casterlevel the level of caster
-	* @returns {int_or_""} the spell range
-	*/
-	findSpellRange = function (customRangeVal, rangeDropdown, casterlevel) {
-		var newRange = 0,
-		ranges = PFUtils.calculateSpellRanges(casterlevel);
-		casterlevel = casterlevel || 0;
-		rangeDropdown=rangeDropdown||'blank';
-		if (rangeDropdown[0] === "{") {
-			rangeDropdown = rangeDropdown.slice(2, rangeDropdown.indexOf("="));
-		}
-		//TAS.debug("at find SpellRange. rangetext:"+customRangeVal +", rangeDropdown:"+rangeDropdown+", area:"+area+", casterlevel:"+casterlevel);
-		switch (rangeDropdown) {
-			case 'number':
-			case 'custom':
-				newRange = parseInt(customRangeVal, 10) || 0;
-				break;
-			case 'perlevel':
-				newRange = (parseInt(customRangeVal, 10) || 0) * casterlevel;
-				break;
-			case 'close':
-				newRange = ranges.close;
-				break;
-			case 'medium':
-				newRange = ranges.medium;
-				break;
-			case 'long':
-				newRange = ranges["long"];
-				break;
-			case 'see text':
-			case 'touch':
-			case 'personal':
-			case 'blank':
-			default:
-				newRange = 0;
-				break;
-		}
-		//TAS.debug("returning customRangeVal "+newRange+" for "+rangeDropdown);
-		return newRange;
-	},
-	/** getWoundPenalty - applies Endurance feat or Gritty Mode to wound level.
-	*@param {int} woundLevel value of wounds attribute
-	*@param {bool} hasEndurance if char has Endurance feat (lessens penalty by 1)
-	*@param {bool} grittyMode if using grittyMode (doubles penalty, applied before hasEndurance)
-	*@returns {int} value to apply.
-	*/
-	getWoundPenalty = function (woundLevel, hasEndurance, grittyMode) {
-		return (woundLevel !== 0) ? (-1 * ((woundLevel * (grittyMode + 1)) - hasEndurance)) : 0;
-	},
-	/** getRepeatingIDStr - if id is not empty, then returns the ID with an underscore on the right. else returns empty string
-	* this is used so the same function can be written for loops from getIDs or direct from the event with no ID
-	*@param {string} id the id of the row or blank
-	*@returns {string} id_  or blank
-	*/
-	getRepeatingIDStr = function (id) {
-		return SWUtils.getRepeatingIDStr(id);
-	},
-	isOptionTemplateReversed = function (spellOptionKey) {
-		return spellOptionKey === "range_pick";
-	},
-	/** getOptionsCompiledRegexMap - finds {{key=*}} in a string to search rolltemplate macros
-	* uses lookahead and lookbehind  to ensure must be preceeded by start or }} , followed by end or {{ 
-	* @param {jsobj map} options map {} of key , only key looked at.
-	* @returns {jsobj map} of key to "{{key=*}}" but as a compiled regex 
-	*/
-	getOptionsCompiledRegexMap = function (options) {
-		return _.mapObject(options, function (outputstr, key) {
-			if (!isOptionTemplateReversed(key)) {
-				return new RegExp("\\s*((?=\\{\\{)|(?=^))\\{\\{" + key + "\\=.*?\\}\\}\\s*((?=\\{\\{)|(?=$))");
-			}
-			return new RegExp("((?=\\{\\{)|(?=^))\\s*\\{\\{\\.*?\\=" + key + "\\}\\}\\s*((?=\\{\\{)|(?=$))");
-		});
-	},
-	/** shouldNotDisplayOption- returns true if the value is the default so we know not to bother displaying in roll.
-	* @param {string} attr: can pass either the attribute or the option name it will be sent to
-	* @param {string} val : the value of the attribute
-	* @returns {boolean}
-	*/
-	shouldNotDisplayOption = function (attr, val) {
-		if (!val) {
-			return true;
-		}
-		switch (attr) {
-			case 'sr':
-				return (!(/^y/i.test(val)));
-			case 'save':
-			case 'saving_throw':
-				return ((/^n/i.test(val) || /harmless/i.test(val)) && !(/and|or/i.test(val)));
-			case 'spell_fail':
-				return (! ( (parseInt(val,10)||0) === 0));
-			default:
-				return false;
-		}
-	},
-	/** deleteOption - removes option text from string and adds {{optionKey=}}
-	* @param {string} optionText the string of a rolltemplate macro
-	* @param {string} optionKey the key from rolltemplate setting, as in: {{optionKey=xxxx}}
-	* @param {string} regexMap output of keys, what to search for from getOptionsCompiledRegexMap()
-	* @returns {string} optionText with the optionKey portion of macro replaced with empty value
-	*/
-	deleteOption = function (optionText, optionKey, regexMap) {
-		var repStr = PFUtils.isOptionTemplateReversed(optionKey) ? "{{=" + optionKey + "}}" : "{{" + optionKey + "=}}";
-		//TAS.debug("deleteOption optionKey"+optionKey+", regexMap[optionKey]:"+regexMap[optionKey]+", repStr:"+repStr);
-		if (optionKey && optionText && regexMap[optionKey]) {
-			optionText = optionText.replace(regexMap[optionKey], repStr);
-		}
-		return optionText;
-	},
-	/**getAvgHP returns average hp for given hit dice and die
-	* also can return 75% or 100% of max hp
-	* @param {int} hdice # of dice
-	* @param {int} hdie # of sides (4,6,8,10,12,etc)
-	* @param {float} mult optional percent of max to average, must be .5 (average), .75, or 1. If null then assume .5
-	* @param {bool} firstMax if true then 1st level gets 100% hp
-	* @param {bool} ispfs if true then round up EVERY level.
-	* @returns {int} hit point average. 
-	*/
-	getAvgHP = function (hdice, hdie, mult, firstMax, ispfs) {
-		var hp=0, bonus=1;
-		//TAS.debug("PFUtils.getAvgHP called with hdice:"+hdice+", hdie:"+hdie+", mult:"+mult+", firstMax:"+firstMax);
-		if (!(mult === 0.5 || mult === 0.75 || mult === 1)) {
-			mult = 0.5;
-		}
-		if (ispfs) { 
-			bonus = 2;
-			mult = 0.5;
-		}
-		if (mult === 1) {
-			hp = hdie * hdice;
-		} else {
-			if (firstMax) {
-				hdice --;
-			}
-			hp= Math.floor( (100*(hdie + bonus) * mult * hdice)/100);
-			if (firstMax){
-				hp+=hdie;
-			}
-		}
-		return hp;
-	},
-	/** takes value of auto hit point radio and returns percent it represents 50,75,100.
-	*@param {int} autohp_percent the value of attr_autohp_percent 
-	*@returns {decimal} either 0.5, 0.75,  or 1.00
-	*/
-	getAutoHPPercentMultiplier = function (autohp_percent) {
-		var  newhealth=0;
-		autohp_percent = parseInt(autohp_percent,10)||0;
-		switch (autohp_percent){
-			case 1: newhealth=0.5;  break;
-			case 2: newhealth=0.75; break;
-			case 3: newhealth=1;   break;
-			default: newhealth=0.5; break;
-		}
-		//TAS.debug("at getAutoHPPercentMultiplier called with "+autohp_percent+", returning with :" + newhealth);
-		return newhealth;
-	},
-	/** parseSpellRangeText - Initial parse of a string from spell , it returns the value to set in the dropdown,
-	* plus whether to run the range text through a secondary parse for numbers.
-	* returns object with keys: dropdown, useorig, number, rangetext
-	* (number only returned if number is a flat number)
-	* @param {string} range the range string from a spell
-	* @param {string} area the area or target string from a spell (whichever filled in, only 1 will be)
-	* @returns {jsobj} map format: {"dropdown":newRangeDropdown,"useorig":useOrigRangeText if special,"number":flatRange,"rangetext":newRangeText if we need to fill in text}
-	*/
-	parseSpellRangeText = function (range, area) {
-		var newRangeDropdown = "",
-		tempRange = 0,
-		tempMatches,
-		tempMatches2,
-		useOrigRangeText = false,
-		flatRange = -1,
-		areaRange,
-		newRangeText = "";
-		//TAS.debug("at parseSpellRangeText: range:"+range+", area:"+area);
-		try {
-			if (!range) {
-				//if range is blank, use the number in area/effect/targets since it will be "30ft emanation" or something similar
-				if (!area) {
-					return {
-						"dropdown": "blank",
-						"useorig": false,
-						"rangetext": "",
-						"number": 0
-					};
-				}
-				areaRange = parseSpellRangeText(area, null);
-				if (areaRange.dropdown === "unknownrange") {
-					areaRange.dropdown = "blank";
-				}
-				if (!(areaRange.dropdown === "number" || areaRange.dropdown === "perlevel")) {
-					areaRange.useorig = false;
-					areaRange.rangetext = "";
-				}
-				return areaRange;
-			}
-			//begin
-			range = range.toLowerCase();
-			//if unlimited use area/target field
-			if (!newRangeDropdown && range === "unlimited") {
-				areaRange = parseSpellRangeText(area, null);
-				if (areaRange.dropdown === "unknownrange") {
-					newRangeDropdown = "blank";
-				} else {
-					newRangeDropdown = areaRange.dropdown;
-					if (!/short|medium|long/.test(newRangeDropdown)) {
-						useOrigRangeText = areaRange.useorig;
-						if (useOrigRangeText && areaRange.rangetext) {
-							range = areaRange.rangetext;
-						}
-					}
-				}
-			}
-			if (!newRangeDropdown) {
-				//and or or - use the value after and/or if there is one, and keep rangetext
-				tempMatches = range.match(/(.*?)\s+(or|and)\s+/);
-				if (tempMatches && tempMatches[1]) {
-					areaRange = parseSpellRangeText(range.substring(tempMatches[0].length), null);
-					if (areaRange && (!(areaRange.dropdown === "unknownrange" || areaRange.dropdown === "blank"))) {
-						newRangeDropdown = areaRange.dropdown;
-						if (areaRange.rangetext) {
-							//If second value is a flat number or per level
-							// then move it BEFORE the and/or so parseInt on rangetext works.
-							if (newRangeDropdown === "number") {
-								if (areaRange.rangetext) {
-									range = areaRange.rangetext + " " + tempMatches[2] + " " + tempMatches[1];
-								} else {
-									range = areaRange.number + " ft. " + tempMatches[2] + " " + tempMatches[1];
-								}
-							} else if (newRangeDropdown === "perlevel") {
-								//must add /level when it is and/or but otherwise not.
-								if (areaRange.rangetext) {
-									range = areaRange.rangetext + "/level " + tempMatches[2] + " " + tempMatches[1];
-								} else {
-									range = areaRange.number + "ft. /level  " + tempMatches[2] + " " + tempMatches[1];
-								}
-							} else {
-								range = tempMatches[1] + " " + tempMatches[2] + " " + areaRange.rangetext;
-							}
-						}
-						useOrigRangeText = true;
-					}
-				}
-			}
-			if (!newRangeDropdown) {
-				if (range === "you") {
-					newRangeDropdown = "personal";
-				} else {
-					tempMatches = range.match(/close|short|medium|long|touch|see text|personal|special|\/level/);
-					if (tempMatches && tempMatches[0]) {
-						switch (tempMatches[0]) {
-							case 'close':
-							case 'medium':
-							case 'long':
-							case 'personal':
-							case 'touch':
-							case 'see text':
-								newRangeDropdown = tempMatches[0];
-								break;
-							case 'short':
-								newRangeDropdown = "close";
-								break;
-							case 'special':
-								newRangeDropdown = "see_text";
-								break;
-							case '/level':
-								tempMatches2 = range.match(/(\d+)(\D*)\/level/);
-								if (tempMatches2 && tempMatches2[1]) {
-									tempRange = parseInt(tempMatches2[1], 10) || 0;
-									range = tempMatches2[1] + (tempMatches2[2] || "");
-									useOrigRangeText = true;
-									newRangeDropdown = "perlevel";
-									flatRange = tempRange;
-								}
-								break;
-						}
-					}
-				}
-			}
-			if (!newRangeDropdown) {
-				//number in front usually emanation, line, cone, etc
-				tempRange = parseInt(range, 10);
-				if (!isNaN(tempRange) && tempRange > 0) {
-					newRangeDropdown = "number";
-					flatRange = tempRange;
-					useOrigRangeText = true;
-				} else {
-					//number in middle after "more than" or "within"
-					tempMatches2 = range.match(/within\s|more\sthan\s/);
-					if (tempMatches2 && tempMatches2[0]) {
-						range = range.substring(tempMatches2[0].index + tempMatches2[0].length);
-						tempRange = parseInt(range, 10);
-						if (!isNaN(tempRange) && tempRange > 0) {
-							newRangeDropdown = "number";
-							flatRange = tempRange;
-							useOrigRangeText = true;
-						}
-					}
-				}
-			}
-			if (!newRangeDropdown && area) {
-				//give up , retry using the text in area/target/effect
-				areaRange = parseSpellRangeText(area, null);
-				newRangeDropdown = areaRange.dropdown;
-				if (newRangeDropdown === "number" || newRangeDropdown === "perlevel") {
-					useOrigRangeText = true;
-					range = areaRange.rangetext;
-				}
-			}
-		} catch (errorParsing) {
-			TAS.error("parseSpellRangeText, error", errorParsing);
-			newRangeDropdown = "unknownrange";
-			useOrigRangeText = true;
-		}
-		if (!newRangeDropdown) {
-			newRangeDropdown = "unknownrange";
-			useOrigRangeText = true;
-		}
-		if (useOrigRangeText === true) {
-			if (newRangeDropdown !== "unknownrange") {
-				//erase everything in parenthesis - also ltrim and rtrim
-				newRangeText = range.replace(/\s*\(.*?\)/, '').replace(/^\s+/, '').replace(/\s+$/, '').replace('feet', 'ft.');
-			} else {
-				newRangeText = range;
-			}
-		}
-		return {
-			"dropdown": newRangeDropdown,
-			"useorig": useOrigRangeText,
-			"number": flatRange,
-			"rangetext": newRangeText
-		};
-	},
-	/** parseCost gets cost in gp 
-	*@param {string} str the string containing the cost: 35gp, 20sp, etc 
-	*@returns {int} cost in gp.
-	*/
-	getCostInGP = function (str){
-		var temp=0,
-		matches = str.match(/(\d+)/);
-		if (matches) { 
-			temp = parseInt(matches[1],10)||0;
-			matches = str.match(/(gp|cp|sp|pp)/i);
-			if (matches){
-				switch(matches[1]){
-					case 'pp':
-						temp = temp*10;
-						break;
-					case 'sp':
-						temp = temp / 10;
-						break;
-					case 'cp':
-						temp = temp / 100;
-						break;
-				}
-			}
-		}
-		return temp;
-	},
-	getIntFromString= function(str){
-		var temp=0, sign=1, matches;
-		matches = PFConst.minusreg.exec(str);
-		if(matches){
-			sign=-1;
-			str = str.replace(matches[0],'');
-		}
-		matches = str.match(/(\d+)/);
-		if (matches) { 
-			temp = sign * (parseInt(matches[1],10)||0);
-		}
-		return temp;		
-	},
-	getCritFromString =function(str){
-		var ret={'crit':20,'critmult':2},matches;
-		matches = PFConst.critreg.exec(str);
-		//TAS.debug("at getCritFromString:"+str+", matches:",matches);
-		if (matches){
-			ret.crit = matches[1];
-			ret.critmult=matches[2];
-		}
-		return ret;
-	},
-	getDiceDieFromString = function(str){
-		var matches,ret={'dice':0,'die':0,'plus':0},sign=1;
-		matches = PFConst.diceDiereg.exec(str);
-		if (matches){
-			ret.dice=parseInt(matches[1],10)||0;
-			ret.die=parseInt(matches[2],10)||0;
-			try {
-				if (matches[3] && PFConst.minusreg.test(matches[3])){
-					sign=-1;
-				}
-				if (matches[4]){
-					ret.plus = sign * (parseInt(matches[4],10)||0);
-				}
-			} catch (err){
-				TAS.error("getDiceDieFromString error finding plus ",err);
-			}
-		}
-		//TAS.debug("at getDiceDieFromString parsing "+str,matches);
-		return ret;
-	},
-	getSpecialAbilityTypeFromString = function(str){
-		var ret='',matches;
-		if(!str){return '';}
-		matches = (/\b(Su|Ex|Sp)\b/i).exec(str);
-		if (matches){
-			return matches[1][0].toUpperCase()+matches[1][2].toLowerCase();
-		}
-		return '';
-	},
-	/** returns string after first comma ( that is after an opening parenthesis )
-	* or after first comma if there is no opening parenthesis
-	* @param {string} str the string to split
-	* @param {bool} putOutside if true then return whateever is before first comma and after opening paren.
-	*      if false, then return everything up to first paren then between 1st and 2nd comma. why? who the hell knows?
-	* @returns {?} ?
-	*/
-	removeUptoFirstComma = function (str, putOutside) {
-		var parensplit,
-		commasplit,
-		retstr,
-		i;
-		if (str.indexOf('(') < 0 || str.indexOf(',') < 0) {
-			return str;
-		}
-		parensplit = str.split(/\s*\(\s*/);
-		if (parensplit.length>1){
-			commasplit = parensplit[1].split(/,\s*/);
-		} else {
-			commasplit = str.split(/,\s*/);
-		}
-		retstr = putOutside ? commasplit[0] : parensplit[0] + '(' + commasplit[1];
-		//rejoin rest of string this is really slow, why bother doing it this way?
-		if (commasplit.length > 2) {
-			for (i = 2; i < commasplit.length; i++) {
-				retstr += ',' + commasplit[i];
-			}
-		}
-		return retstr;
-	},
-	/**replaceDiceDieString puts inline roll brackets [[ ]] around 'xdy +z' dice strings (z exists or not)
-	*@param {string} str a string which includes a diceroll substring xdy or xdy +/-z
-	*@returns {string} same string with brackets around dice roll
-	*/
-	replaceDiceDieString = function (str) {
-		var tempstr = "",
-		tempstrs = str.split(/(\d+d\d+\s*[+\-]??\d*)/i);
-		tempstr = _.reduce(tempstrs, function (memo, splitted) {
-			//TAS.debug('at ' + splitted);
-			if ((/(\d+d\d+\s*[+\-]??\d*)/i).test(splitted)) {
-				return memo + "(" + splitted + "): [[" + splitted + "]] ";
-			}
-			return memo + splitted;
-		}, "");
-		return tempstr;
-	},
-	getDiceDieString = function(str){
-		var reg=/(\d+d\d+\s*[+\-]??\d*)/i,
-		matches;
-		matches=reg.exec(str);
-		if(matches){
-			return "[["+ matches[0] + "]]";
-		}
-		return "";
-	},
-	/**getDCString - gets macro formula for special ability calculating DC using ability score, what the level attribute is, and 
-	* whether to divide that level by 2 or not.
-	* @param {string} ability the ability score string the DC is based on. Usually CON for special abilities.
-	* @param {string} levelAttr optional the level attribute , either "level" or "class-0-level" or "npc-hd-num" etc
-	* @param {bool} isUndead flag if undead, if true, then if ability is 'CON' change to 'CHA'
-	* @param {int} miscBonus a flat number to add in
-	* @param {bool} doNotDivideByTwo if true then do not divide level attr value by 2 
-	* @returns {string} default is: "DC [[ 10 + @{" + ability + "-mod} + floor(@{"+levelAttr+"}/2) ]]";
-	*/
-	getDCString = function (ability, levelAttr, isUndead, miscBonus, doNotDivideByTwo) {
-		var tempstr = '', pre = 'floor(', post = '/2)';
-
-		tempstr = "DC [[ 10 ";
-		if (ability) {
-			if (isUndead && ability === 'CON') {
-				ability = 'CHA';
-			}
-			tempstr += "+ @{" + ability + "-mod} ";
-		}
-		if (levelAttr) {
-			if (doNotDivideByTwo) {
-				pre = ''; post = '';
-			} else {
-				tempstr += "+ " + pre + "@{" + levelAttr + "}" + post + " ";
-			}
-		}
-		if(miscBonus) {
-			tempstr +=  "+ " + miscBonus ;
-		}
-		tempstr += " ]]";
-
-		return tempstr;
-	},
-	/**replaceDCString looks for DC n, and replaces "n" with the [[ calculated DC  ]] by calling getDCString
-	* @param {string} str the string to search and replace
-	* @param {string} ability the ability score string the DC is based on. Usually CON for special abilities.
-	* @param {string} levelAttr optional the level attribute , either "level" or "class-0-level" or "npc-hd" etc
-	* @param {bool} isUndead flag if undead, if true, then if ability is 'CON' change to 'CHA'
-	* @param {int} levelFlatNum optional the level, if levelAttr is blank, this must be filled in, or vice versa
-	* @param {bool} doNotDivideByTwo if true then do not divide level by 2 to calculate DC
-	* @returns {string} default is: "DC [[ 10 + @{" + ability + "-mod} + floor(@{"+levelAttr+"}/2) ]]"
-	*/
-	replaceDCString = function (str, ability, levelAttr, isUndead, levelFlatNum, doNotDivideByTwo) {
-		var tempstr = '', matches,pre='',post='', retstr=str,rawDC=10;
-		try {
-			matches = str.match(/D[Cc]\s*\d+/);
-			if (matches){
-				tempstr =matches[0].match(/\d+/);
-				rawDC=parseInt(tempstr,10)||0;
-				tempstr = getDCString(ability, levelAttr, isUndead, levelFlatNum, doNotDivideByTwo);
-				pre= str.slice(0, matches.index)||'';
-				post = str.slice(matches.index + matches[0].length)||'';
-				retstr=pre + tempstr + post;
-			}
-		} catch(er) {
-			TAS.error("at replaceDCString, cannot find DC string in "+ str,er);
-		} finally {
-			return retstr;
-		}
-	},
-	/** returns rest of string after number 
-	*@param {string} str the string
-	*@returns {string} rest of string after finding a number.
-	*/
-	getNoteAfterNumber = function (str) {
-		str = str.slice(str.indexOf(/\d+/));
-		return str;
-	},
-	/**gets value 'field_compendium' from v,passes it to synchronous methodToCall mapping function, then sets in 'field' 
-	*@param {string} prefix the repeating_section_id_  string
-	*@param {string} field the name of compendium field , must have _compendium at end. Without '_compendium' this is the write field
-	*@param {function} methodToCall synchronous function that maps value of field_compendium to another val to set
-	*@param {jsmap} v the values returned from getAttrs
-	*@param {jsmap} setter to pass to setAttrs
-	*@param {string} setField optional if the attr to write to is not 'field' it will be prefix+setField
-	*/
-	getCompendiumFunctionSet = function (prefix,field,methodToCall,v,setter,setField){
-		var temp=0,
-			attr=v[prefix+field+'_compendium'];
-		if (attr){
-			temp= methodToCall(attr);
-			if (temp) { 
-				setField=setField||field;
-				setter[prefix+field]= temp;
-			}
-		}
-		return setter;
-	},
-	/**gets int value 'field_compendium' from v, then sets in 'field' 
-	*@param {string} prefix the repeating_section_id_  string
-	*@param {string} field the name of compendium field , must have _compendium at end. Without '_compendium' this is the write field
-	*@param {jsmap} v the values returned from getAttrs
-	*@param {jsmap} setter to pass to setAttrs
-	*@param {string} setField optional if the attr to write to is not 'field' it will be prefix+setField
-	*/
-	getCompendiumIntSet = function (prefix,field,v,setter,setField){
-		var tempInt=0,attr;
-		try {
-			attr=v[prefix+field+'_compendium'];
-			if (attr){
-				tempInt= getIntFromString(attr);
-				//TAS.debug("get int field:"+field+", val="+attr+", int:"+tempInt);
-				if (tempInt) { 
-					setField=setField||field;
-					setter[prefix+field]= tempInt;
-				}
-			}
-		} catch (err){
-			TAS.error("getCompendiumIntSet error on :"+prefix+", field:" +field + ", setField:"+setField,err);
-		} finally {
-			return setter;
-		}
-	},
-	getRowId = function(sourceAttribute){
-		return SWUtils.getRowId(sourceAttribute);
-	},
-	getAttributeName = function(source){
-		return SWUtils.getAttributeName(source);
-	},
-	removeWhisperFromMacro = function(macrostr){
-		var matches;
-		if(!macrostr) {return macrostr;}
-		//use hisper since some have capital W others not
-		matches = macrostr.match(/whisper\}/i);
-		if(matches){
-			return SWUtils.trimBoth(macrostr.slice(matches.index+matches[0].length));
-		}
-		return macrostr;
-	}
-	;
-	console.log(PFLog.l + '   PFUtils module loaded          ' + PFLog.r, PFLog.bg);
-	PFLog.modulecount++;
-	return {
-		shouldNotDisplayOption: shouldNotDisplayOption,
-		deleteOption: deleteOption,
-		isOptionTemplateReversed: isOptionTemplateReversed,
-		getOptionsCompiledRegexMap: getOptionsCompiledRegexMap,
-		findAbilityInString: findAbilityInString,
-		getRepeatingIDStr: getRepeatingIDStr,
-		calculateSpellRanges: calculateSpellRanges,
-		findSpellRange: findSpellRange,
-		parseSpellRangeText: parseSpellRangeText,
-		removeUptoFirstComma: removeUptoFirstComma,
-		replaceDiceDieString: replaceDiceDieString,
-		getDiceDieString: getDiceDieString,
-		getDCString: getDCString,
-		replaceDCString: replaceDCString,
-		getNoteAfterNumber: getNoteAfterNumber,
-		getWoundPenalty: getWoundPenalty,
-		getAutoHPPercentMultiplier: getAutoHPPercentMultiplier,
-		getAvgHP: getAvgHP,
-		getCostInGP: getCostInGP,
-		getCompendiumFunctionSet: getCompendiumFunctionSet,
-		getCompendiumIntSet: getCompendiumIntSet,
-		getIntFromString: getIntFromString,
-		getCritFromString: getCritFromString,
-		getSpecialAbilityTypeFromString: getSpecialAbilityTypeFromString,
-		getDiceDieFromString: getDiceDieFromString,
-		getRowId: getRowId,
-		getAttributeName: getAttributeName,
-		removeWhisperFromMacro: removeWhisperFromMacro
-	};
-}());
-var PFUtilsAsync = PFUtilsAsync || (function () {
-	'use strict';
-	/****************************ASYNCRHOUNOUS UTILITIES ***********************************
-	***************************************************************************************/
-	/* setDropdownValue
-	* Looks at a dropdown selected value, finds the matching attribute value, and then
-	* sets the writeFields with that number.
-	*
-	* @readField {string} = the dropdpown field
-	* @writeFields {string or Array} = One string or an array of strings that are fields to write the value to
-	* @callback {function} optional = if we need to update the field, call this function
-	*       callback(newvalue,oldvalue,ischanged)
-	*  If writeField is a string not an Array, then set old value as 2nd param (could be NaN)
-	*/
-	var setDropdownValue = function (readField, writeFields, callback, silently) {
-		SWUtils.setDropdownValue(readField, writeFields, PFUtils.findAbilityInString, callback, silently);
-	},
-	/** calls setDropdownValue for a dropdown in a repeating section
-	*@param {string} section the string between "repeating_" and "_<id>"
-	*@param {string} id optional- the id of this row, blank if in context of the current row 
-	*@param {string} from the attribute name of the dropdown , string after "repeating_section_id_"
-	*@param {string} to the attribute to write to, string after "repeating_section_id_"
-	*@param {function} callback - the function passed to setDropdownValue as its callback, that function calls it
-	*/
-	setRepeatingDropdownValue = function (section, id, from, to, callback,silently) {
-		var idStr = PFUtils.getRepeatingIDStr(id),
-		prefix = "repeating_" + section + "_" + idStr;
-		setDropdownValue(prefix + from, prefix + to, callback,silently);
-	},
-	/** setRowIds
-	* sets the ID fields and new_flag fields for all rows in the section
-	* @param {string} section  = the fieldset name after "section_"
-	*/
-	setRowIds = function (section) {
-		getSectionIDs("repeating_" + section, function (ids) {
-			var setter = {};
-			_.each(ids, function (id) {
-				setter["repeating_" + section + "_" + id + "_row_id"] = id;
-			});
-			setAttrs(setter);
-		});
-	},
-	registerEventHandlers = function() {
-		//REPEATING SECTIONS set IDs
-		_.each(PFConst.repeatingSections, function (section) {
-			var eventToWatch = "change:repeating_" + section + ":ids-show";
-			on(eventToWatch, TAS.callback(function eventCheckIsNewRow(eventInfo) {
-				var setter={},id;
-				TAS.debug("caught " + eventInfo.sourceAttribute + " event: " + eventInfo.sourceType);
-				if (eventInfo.sourceType === "player" || eventInfo.sourceType === "api") {
-					id = SWUtils.getRowId(eventInfo.sourceAttribute);
-					setter["repeating_" + section + "_"+id+"_row_id"]=id;
-					setAttrs(setter,PFConst.silentParams);
-				}
-			}));
-		});		
-	}
-	;
-	registerEventHandlers();
-	console.log(PFLog.l + '   PFUtilsAsync module loaded     ' + PFLog.r, PFLog.bg);
-	PFLog.modulecount++;
-	return {
-		setDropdownValue: setDropdownValue,
-		setRepeatingDropdownValue: setRepeatingDropdownValue,
-		setRowIds: setRowIds
-	};
-}());
-var PFMacros = PFMacros || (function () {
-	'use strict';
-	var 
-	/** splitMacro Splits macro into {{x=y}} components
-	 * and @{attr} if at top level (not inside a {{x=@{attr}}})
-	 *@param {string} macrostr the macro-text from a repeating row
-	 *@returns {Array} of strings comprising macro
-	 */
-	splitMacro = function(macrostr){
-		var splitted,newsplit,lastclosing;
-		if (!macrostr) {return "";}
-		splitted = macrostr.split(/(?=\{\{)/);
-		splitted = SWUtils.trimBoth(splitted);
-		newsplit = _.reduce(splitted,function(memo,val){
-			try {
-				if (val.slice(0,2)==='{{') {
-					if (val.slice(-2)==='}}'){
-						memo.push(val);
-					} else {
-						lastclosing= val.lastIndexOf('}}');
-						if (lastclosing <0){
-							TAS.error ("error! no closing brackets for ",val);
-							//just fix it
-							val += '}}';
-							lastclosing= val.lastIndexOf('}}');
-							memo.push(val);
-						} else {
-							memo.push(SWUtils.trimBoth(val.slice(0,lastclosing+2)));
-							memo=memo.concat(SWUtils.trimBoth(SWUtils.trimBoth(val.slice(lastclosing+2)).replace('&amp;','&').split(/(?=[\@\&]\{)/)));
-						}
-					}
-				} else {
-					val=val.replace('&amp;','&'); 
-					memo=memo.concat(SWUtils.trimBoth(val.split(/(?=[\@\&]\{)/)));
-				}
-			} catch (err){
-				TAS.error("splitmacro",err);
-			} finally {
-				return memo;
-			}
-		},[]);
-		return newsplit;
-	},
-	/** arrayToMap Splits array of {{x=y}} to mapping of '{{x=': 'y}}' 
-	* and splits &{template:templatename} on the :
-	* unless the item has no equals sign then the value = map.
-	*    e.g. @{option}  is returned as @{option}=@{option}
-	*@param {Array} currArray of strings for rolltemplate {{key=value}}
-	*@returns {jsobj} of each array entry split in half
-	*/
-	arrayToMap = function(currArray,removeWhisper){
-		return _.reduce(currArray,function(memo,val){
-			var spliteq=val.split('=');
-			if (val){
-				if(spliteq.length===2){
-					memo[spliteq[0]+'=']=spliteq[1];
-				} else if(spliteq.length>2){
-					memo[spliteq[0]+'=']= _.rest(spliteq,1).join('=');
-				} else if ((/template:/).test(val)){
-					memo['&{template:']=val.slice(val.indexOf(':')+1);
-				} else if (!(removeWhisper && (/whisper/i).test(val))){
-					memo[val]=val;
-				}
-			}
-			return memo;
-		},{});
-	},
-	/**mergeMacroMaps merges currMap into defaultMap 
-	*@param {jsobj} currMap map of the current macro text on a sheet created by arrayToMap:
-								{rollqueryleft: rollqueryright}
-	*@param {jsobj} defaultMap {rollqueryleft : {
-								current:rollqueryright, 
-								old:[  oldrollqueryright1, oldrollqueryright2 ], 
-								replacements:[  { from: fromstring, to:tostring}, {from:fromstring, to:tostring}] 
-								}
-							}
-	*@param {Array} sameAsKeys array of strings of keys in defaultMap where value.current is the same string as the key
-	*@returns {Array} of strings for macro entries.
-	*/
-	mergeMacroMaps = function(currMap,defaultMap,sameAsKeys){
-		var currKeys=[],newKeys=[],customKeys=[], newArray=[] , customizedMap={}, userDefinedMap={};
-		try {
-			currKeys = _.keys(currMap).sort();
-			customKeys = _.difference(currKeys,_.keys(defaultMap)).sort();
-			if (!sameAsKeys || _.size(sameAsKeys)===0 ){
-				sameAsKeys = _.reduce(defaultMap,function(memo,val,key){
-					if (val.current===key){
-						memo.push(key);
-					}
-					return memo;
-				},[]).sort();
-			}
-			//TAS.debug("at mergeMacroMaps comparing: ",currMap,defaultMap,sameAsKeys);
-			customizedMap = _.chain(defaultMap)
-				.pick(function(compareobj,defaultKey){
-					//intersection
-					return (_.indexOf(currKeys,defaultKey,true) >=0 );
-				})
-				.omit(function(compareobj,defaultKey){
-					//difference
-					return (_.indexOf(sameAsKeys,defaultKey,true)>=0);
-				})
-				.omit(function(compareobj,defaultKey){
-					//user val = new
-					return (compareobj.current === currMap[defaultKey]);
-				})
-				.omit(function(compareobj,defaultKey){
-					//user val = one of old vals
-					return ( !(compareobj.old) || (_.indexOf(compareobj.old,currMap[defaultKey]) >=0)  );
-				})
-				.mapObject(function(compareobj,defaultKey){
-					//only customized values left, if any
-					var newString='';
-					try {
-						newString = currMap[defaultKey];
-						if (newString){
-							if (compareobj.replacements){
-								newString = _.reduce(compareobj.replacements,function(memo,replacer){
-									return memo.replace(replacer.from,replacer.to);
-									},newString);		
-							}
-							newString = defaultKey + newString;
-						}
-					} catch (erri) {
-						TAS.error('mergeMacroMaps  erri on '+defaultKey,erri);
-					} finally {
-						return newString;
-					}
-				}).value();
-
-			//extra stuff from user.
-			userDefinedMap = _.reduce (customKeys,function(memo,currKey){
-					if( currKey !== currMap[currKey] ){
-						memo[currKey] = currKey+currMap[currKey];
-					} else {
-						memo[currKey] = currKey;
-					}
-					return memo;
-				},{});
-				
-			//TAS.info("joining ",defaultMap,customizedMap,userDefinedMap);
-			//array in order of defaultMap, add in custom values, then turn to array
-			newArray = _.chain(defaultMap)
-				.mapObject(function(compareobj,defaultKey){
-					if (_.indexOf(sameAsKeys,defaultKey)>=0){
-					
-						return defaultKey;
-					}
-					return defaultKey + compareobj.current;
-				})
-				.extend(customizedMap)
-				.extend(userDefinedMap)
-				.values()
-				.value();
-
-		} catch(err3) {
-			TAS.error("getNewArray outer error",err3);
-		} finally {
-			return newArray;
-		}
-	},
-	/**migrateMacro makes sure one macro is up to date, synchronous.
-	*@param {string} currMacro current macro from sheet
-	*@param {string} defaultMacro default / new correct macro string
-	*@param {jsobj} defaultMap {rollqueryleft : {
-								current:rollqueryright, 
-								old:[  oldrollqueryright1, oldrollqueryright2 ], 
-								replacements:[  { from: fromstring, to:tostring}, {from:fromstring, to:tostring}] 
-								}
-							}
-	*@param {Array} deleteArray array of strings to just delete from the currMacro.
-	*@param {Array} sameAsKeys array of strings of keys in defaultMap where value.current is the same string as the key
-	*@returns {string} one of 3 values:
-				null if caller should NOT update macro, 
-				"BLANK" if caller should update macro attribute with "" to reset it.
-				any other string: the new macro (if the user customized it, then this is the new one with updates)
-	*/
-	migrateMacro = function (currMacro,defaultMacro,defaultMap,deleteArray,sameAsKeys) {
-		var currMacroArray, currMacroMap,newMacroArray,newMacroString=null;
-		try {
-			if (currMacro !== defaultMacro){
-				if (deleteArray && Array.isArray(deleteArray)){
-					_.each(deleteArray,function(strToDelete){
-						currMacro = currMacro.replace(strToDelete,'');
-					});
-				}
-				if (currMacro===defaultMacro){
-					newMacroString=null;
-				} else {
-					currMacroArray = splitMacro(currMacro);
-					currMacroArray = _.reject(currMacroArray,function(val){
-						return (!val || val==="0" || val==="1" || val.indexOf("undefined")>=0);
-					});
-					currMacroMap = arrayToMap(currMacroArray,true);
-					//TAS.info("migrateMacro calling map with ",currMacroMap,defaultMap,sameAsKeys);
-					newMacroArray = mergeMacroMaps (currMacroMap,defaultMap,sameAsKeys);
-					//TAS.debug("migrateMacro received back ",newMacroArray);
-					newMacroArray = _.reject(newMacroArray,function(val){
-						return (!val || val==="0" || val==="1" || val.indexOf("undefined")>=0);
-					});
-					newMacroString = newMacroArray.join(' ');
-					if (newMacroString === defaultMacro){
-						newMacroString= 'BLANK';
-					}
-				}
-			}
-		} catch (err){
-			TAS.error("PFMacros.migrateMacro error on "+currMacro,err);
-		} finally {
-			return newMacroString;
-		}
-	},
-	/**migrateRepeatingMacros updates all macros in the section 
-	*@param {function} callback after calling setAttrs with the new macros
-	*@param {string} section  name after "repeating_"
-	*@param {string} fieldname  the attribute name containing the macro after "id_"
-	*@param {string} defaultMacro the current Macro in the page
-	*@param {jsobj} defaultMap map of "{{rolltemplatekey=" to right side "var}}"
-			{rollqueryleft : {
-				current:rollqueryright, 
-				old:[  oldrollqueryright1, oldrollqueryright2 ], 
-				replacements:[  { from: fromstring, to:tostring}, {from:fromstring, to:tostring}] 
-				}
-			}
-	*@param {Array} deleteArray  array of strings of old rolltemplate entries that are not used (entire entry not just left side )
-	*/
-	migrateRepeatingMacros = function (callback,section,fieldname,defaultMacro,defaultMap,deleteArray, whisper){
-		var done = _.once(function(){
-			TAS.debug("leaving migrateRepeatingMacros for "+ section + ", "+fieldname);
-			if (typeof callback === "function") {
-				callback();
-			}
-		});
-		whisper=whisper||'';
-		if (whisper) {whisper+=' ';}
-		getSectionIDs('repeating_'+section,function(ids){
-			var fields=[],prefix='repeating_'+section+'_';
-			if (!ids || _.size(ids)===0){
-				done();
-				return;
-			}
-			_.each(ids,function(id){
-				fields.push(prefix+id+'_'+fieldname);
-			});
-			getAttrs(fields,function(v){
-				var setter={},sameAsKeys=[];
-				sameAsKeys = _.reduce(defaultMap,function(memo,val,key){
-					if (val.current===key){
-						memo.push(key);
-					}
-					return memo;
-				},[]).sort();
-				setter=_.reduce(v,function(memo,currMacro,key){
-					var newMacro="";
-					try {
-						if (!currMacro){
-							//setting to blank does not seem to work, it keeps coming back as undefined, so set to default.
-							memo[key]=whisper +defaultMacro;
-						} else {
-							currMacro=PFUtils.removeWhisperFromMacro(currMacro);
-							newMacro = migrateMacro(currMacro,defaultMacro,defaultMap,deleteArray,sameAsKeys);
-							if (newMacro==="BLANK") {
-								memo[key]="";
-							} else if (newMacro){
-								memo[key]=whisper + newMacro;
-							}
-						}
-					} catch (innererr){
-						TAS.error("migrateRepeatingMacros error migrating "+ key,innererr);
-					} finally {
-						return memo;
-					}
-				},{});
-				if (_.size(setter)>0){
-					setAttrs(setter,{},done);
-				} else {
-					done();
-				}
-			});
-		});		
-	},
-	/**Calls migrateRepeatingMacros once for each version of the parameters in replaceArray
-	* each parameter below potentially has the word 'REPLACE' in it, for each element in replaceArray,
-	* replace the word REPLACE with that element.
-	* This is not the most efficient, but it was alot easier than rewriting migrateRepeatingMacros 
-	*@param {function} callback after calling setAttrs with the new macros
-	*@param {string} section  name after "repeating_"
-	*@param {string} fieldname  the attribute name containing the macro after "id_"
-	*@param {string} defaultMacro the current Macro in the page
-	*@param {jsobj} defaultMap map of "{{rolltemplatekey=" to right side "var}}"
-			{rollqueryleft : {
-				current:rollqueryright, 
-				old:[  oldrollqueryright1, oldrollqueryright2 ], 
-				replacements:[  { from: fromstring, to:tostring}, {from:fromstring, to:tostring}] 
-				}
-			}
-	*@param {Array} deleteArray array of strings of old rolltemplate entries that are not used (entire entry not just left side )
-	*@param {Array} replaceArray array of strings to replace the word 'REPLACE' with that are found in the other params.
-	*/
-	migrateRepeatingMacrosMult = function (callback,section,fieldname,defaultMacro,defaultMap,deleteArray,replaceArray, whisper){
-		var done=_.once(function(){
-			TAS.debug("leaving migrateRepeatingMacrosMult for "+section+"_"+fieldname);
-			if (typeof callback === "function"){
-				callback();
-			}
-		}),
-		fieldnames={}, defaultMacros={}, defaultMaps={}, deleteArrays={}, numTimes, doneOnce;
-		try {
-			if(!( replaceArray && Array.isArray(replaceArray) && _.size(replaceArray)>0)){
-				TAS.warn("migrateRepeatingMacrosMult no replace array for "+section+", "+fieldname);
-				done();
-				return;
-			}
-			numTimes = _.size(replaceArray);
-			doneOnce = _.after(numTimes,done);
-			//create new mappings of fieldname, defaultMacro, defaultMap, deleteArray
-			//one new version of each per element in replaceArray.
-			fieldnames = _.reduce(replaceArray,function(m,val){
-				m[val]=fieldname.replace(/REPLACE/g,val);
-				return m;
-			},{});
-			defaultMacros = _.reduce(replaceArray,function(m,val){
-				m[val]=defaultMacro.replace(/REPLACE/g,val);
-				return m;
-			},{});
-			defaultMaps =_.reduce(replaceArray,function(replaceMemo,val){
-				var newMap ={};
-				try {
-					newMap =_.reduce(defaultMap,function(m,currobj,key){
-						var newkey, newobj={};
-						try {
-							newkey = key.replace(/REPLACE/g,val);
-							newobj.current = currobj.current.replace(/REPLACE/g,val);
-							if(currobj.old){
-								newobj.old = _.map(currobj.old,function(oldmacro){
-									return oldmacro.replace(/REPLACE/g,val);
-								});
-							}
-							if (currobj.replacements){
-								newobj.replacements= _.map(currobj.replacements,function(replacementobj){
-									var newreplacement = {};
-									newreplacement.from = replacementobj.from.replace(/REPLACE/g,val);
-									newreplacement.to = replacementobj.to.replace(/REPLACE/g,val);
-									return newreplacement;
-								});
-							}
-							m[newkey]=newobj;
-						} catch (innererr){
-							TAS.error("migrateRepeatingMacrosMult, error creating defaultMaps: replaceval"+val+", key="+key+", matching obj:",currobj);
-						} finally {
-							return m;
-						}
-					},{});
-					replaceMemo[val]=newMap;
-				} catch (errineer2){
-					TAS.error("error building map of defaultMap : replaceval:"+val);
-				} finally {
-					return replaceMemo;
-				}
-			},{});
-			if(deleteArray){
-				deleteArrays = _.reduce(replaceArray,function(m,val){
-					m[val] = _.map(deleteArray,function(delstr){
-							return delstr.replace(/REPLACE/g,val);
-						});
-					return m;
-				},{});
-			}
-			_.each(replaceArray,function(val){
-				var delArray;
-				try {
-					if(fieldnames[val] && defaultMacros[val] && defaultMaps[val] ){
-						delArray= deleteArrays[val]||null;
-						migrateRepeatingMacros(doneOnce,section,fieldnames[val],defaultMacros[val],defaultMaps[val],delArray,whisper);
-					} else {
-						doneOnce();
-					}
-				} catch(err){
-					TAS.error("migrateRepeatingMacrosMult error calling migrateRepeatingMacros for "+val,err);
-					doneOnce();
-				}
-			});
-		} catch (outererr){
-			TAS.ERROR("migrateRepeatingMacrosMult error in outermost section SHOULD NOT HAPPEN ",outererr);
-			done();
-		}
-	},
-	migrateStaticMacro = function(callback, fieldname, defaultMacro, defaultMap, deleteArray, sameAsKeys, whisper){
-		var done = _.once(function(){
-			TAS.debug("leaving migrateRepeatingMacros for "+ fieldname);
-			if (typeof callback === "function") {
-				callback();
-			}
-		});
-		whisper=whisper||'';
-		if (whisper) {whisper+=' ';}
-		//TAS.debug("at PFMacros.migrateStaticMacro for repeating_"+ fieldname);
-		getAttrs([fieldname],function(v){
-			var setter={}, newMacro='', currMacro='';
-			try {
-				if (!sameAsKeys){
-					sameAsKeys = _.reduce(defaultMap,function(memo,val,key){
-						if (val.current===key){
-							memo.push(key);
-						}
-						return memo;
-					},[]).sort();
-				}
-				currMacro = v[fieldname];
-				if(!currMacro){
-					setter[fieldname]=whisper + defaultMacro;
-				} else {
-					currMacro=PFUtils.removeWhisperFromMacro(currMacro);
-					newMacro = migrateMacro(currMacro,defaultMacro,defaultMap,deleteArray,sameAsKeys);
-					if(newMacro){
-						if (newMacro === 'BLANK'){
-							setter[fieldname]="";
-						} else {
-							setter[fieldname]=whisper + newMacro;
-						}
-					}
-				}
-			} catch (innererr){
-				TAS.error("migrateRepeatingMacros error migrating "+fieldname+", "+currMacro,innererr);
-			} finally {
-				if (_.size(setter)>0){
-					setAttrs(setter,{},done);
-				} else {
-					done();
-				}
-			}
-		});
-	},
-	migrateStaticMacros = function (callback,fieldnames,defaultMacros,defaultMaps,deleteArrays, sameAsKeys, whisper){
-		var done = _.once(function(){
-			TAS.debug("leaving migrateStaticMacros ");
-			if (typeof callback === "function") {
-				callback();
-			}
-		}), fields =[], keys=[];
-		whisper=whisper||'';
-		if (whisper) {whisper+=' ';}
-		fields =_.values(fieldnames);
-		keys = _.keys(fieldnames);
-		getAttrs(fields,function(v){
-			var setter={};
-			if (!sameAsKeys || !Array.isArray(sameAsKeys) || _.size(sameAsKeys)===0){
-				sameAsKeys = _.reduce(defaultMaps[0],function(memo,val,key){
-					if (val.current===key){
-						memo.push(key);
-					}
-					return memo;
-				},[]).sort();
-			}
-			setter=_.reduce(keys,function(memo,key){
-				var newMacro="", currMacro="", delArray=null, field='';
-				try {
-					field=fieldnames[key];
-					currMacro = v[field];
-					if(!currMacro){
-						memo[field]=defaultMacros[key];
-					} else {
-						if(deleteArrays && deleteArrays[key]){
-							delArray = deleteArrays[key];
-						}
-						if (currMacro){
-							newMacro = migrateMacro(currMacro,defaultMacros[key],defaultMaps[key],delArray,sameAsKeys);
-							if (newMacro==="BLANK") {
-								memo[field]="";
-							} else if (newMacro){
-								memo[field]=newMacro;
-							}
-						}
-					}
-				} catch (innererr){
-					TAS.error("migrateStaticMacros error migrating "+ key,innererr);
-				} finally {
-					return memo;
-				}
-			},{});
-			//TAS.debug("migrateStaticMacros setting ", setter);
-			if (_.size(setter)>0){
-				setAttrs(setter,{},done);
-			} else {
-				done();
-			}
-		});		
-	},
-	migrateStaticMacrosMult = function(callback, fieldname, defaultMacro, defaultMap, deleteArray, replaceArray, keysToReplaceShortcut, valsToReplaceShortcut, useNoNumber, whisper){
-		var done=_.once(function(){
-			TAS.debug("leaving migrateRepeatingMacrosMult for "+fieldname);
-			if (typeof callback === "function"){
-				callback();
-			}
-		}),
-		fieldnames={}, replacers = {},defaultMacros={}, defaultMaps={}, deleteArrays={}, numTimes, doneOnce, sameAsKeys,
-		replaceStrings = function(str,val){
-			var fromAndTos=[];
-			try {
-				fromAndTos =replacers[val];
-				//TAS.debug("now using replacer to replace :"+val+" in "+str,fromAndTos);
-				if (fromAndTos){
-					_.each(fromAndTos,function(replacer){
-						str = str.replace(replacer.from,replacer.to);
-					});
-				}
-			} catch (err){
-				TAS.error("PFMacros migrateStaticMacrosMult replaceStrings error replacing  "+val+" in "+str,replacers[val],err);
-			} finally {
-				return str;
-			}
-		};
-		try {
-			whisper=whisper||'';
-			if (whisper) {whisper+=' ';}
-			if(!( replaceArray && Array.isArray(replaceArray) )){
-				TAS.warn("migrateStaticMacrosMult no replace array for  "+fieldname);
-				done();
-				return;
-			}
-			numTimes = _.size(replaceArray);
-			doneOnce = _.after(numTimes,done);
-			//create new mappings of fieldname, defaultMacro, defaultMap, deleteArray
-			//one new version of each per element in replaceArray.
-			if (useNoNumber){
-				replacers = _.reduce(replaceArray,function(m,val){
-					var valNoNum = val.replace(/\d+/g,'').replace(/\-+$/,'');
-						m[val] = [{ 'from':/REPLACELOWERREMOVENUMBER/g, 'to': valNoNum.toLowerCase()},
-							{'from':/REPLACELOWER/g, 'to': val.toLowerCase()},
-							{'from':/REPLACEREMOVENUMBER/g, 'to': valNoNum},
-							{'from':/REPLACE/g, 'to': val}];
-						return m;
-				},{});
-			} else {
-				replacers = _.reduce(replaceArray,function(m,val){
-					m[val] = [{'from':/REPLACELOWER/g, 'to': val.toLowerCase()},
-						{'from':/REPLACE/g, 'to': val}];
-					return m;
-				},{});
-			}
-			fieldnames = _.reduce(replaceArray,function(m,val){
-				m[val]=fieldname.replace(/REPLACE/g,val);
-				return m;
-			},{});
-			sameAsKeys = _.reduce(defaultMap,function(memo,compObj,key){
-				var replacedKeys=[];
-				if (compObj.current===key){
-					//see if this needs replacing.
-					if (_.indexOf(keysToReplaceShortcut,key,true)>=0){
-						replacedKeys = _.map(replaceArray,function(replaceKey){
-							return key.replace(/REPLACE/g, replaceKey);
-						});
-						memo = memo.concat(replacedKeys);
-					} else {
-						memo.push(key);
-					}
-				}
-				return memo;
-			},[]).sort();
-
-			defaultMacros = _.reduce(replaceArray,function(m,val){
-				var newMacro = replaceStrings(defaultMacro, val);
-				m[val]=newMacro;
-				return m;
-			},{});
-			defaultMaps =_.reduce(replaceArray,function(replaceMemo,val){
-				var newMap ={};
-				try {
-					newMap =_.reduce(defaultMap,function(memo,currobj,key){
-						var newkey, newobj={};
-						try {
-							if (_.indexOf(keysToReplaceShortcut,key,true)>=0){
-								newkey = key.replace(/REPLACE/g,val);
-							} else {
-								newkey = key;
-							}
-							if (_.indexOf(valsToReplaceShortcut,key,true)>=0){
-								newobj.current = replaceStrings(currobj.current, val);
-								if(currobj.old){
-									newobj.old = _.map(currobj.old,function(oldmacro){
-										return replaceStrings(oldmacro, val);
-									});
-								}
-							} else {
-								newobj.current = currobj.current;
-								if (currobj.old){
-									newobj.old = currobj.old;
-								}
-							}
-							if (currobj.replacements){ 
-								newobj.replacements = currobj.replacements;
-							}
-							memo[newkey]=newobj;
-						} catch (innererr) {
-							TAS.error("migrateStaticMacrosMult error creating defaultMaps: replaceval"+val+", key="+key+", matching obj:",currobj);
-						} finally {
-							return memo;
-						}
-					},{});
-					replaceMemo[val]=newMap;
-				} catch (errineer2){
-					TAS.error("migrateStaticMacrosMult error building map of defaultMap : replaceval:"+val);
-				} finally {
-					return replaceMemo;
-				}
-			},{});
-			
-			if(deleteArray){
-				deleteArrays = _.reduce(replaceArray,function(m,val){
-					m[val] = _.map(deleteArray,function(delstr){
-							return replaceStrings(delstr,val);
-						});
-					return m;
-				},{});
-			}
-			migrateStaticMacros(done,fieldnames,defaultMacros,defaultMaps,deleteArrays,sameAsKeys,whisper);
-		} catch (outererr){
-			TAS.ERROR("migrateStaticMacrosMult error in outermost section SHOULD NOT HAPPEN ",outererr);
-			done();
-		}
-	};
-	console.log(PFLog.l + '   PFMacros module loaded         ' + PFLog.r, PFLog.bg);
-	PFLog.modulecount++;
-	return {
-		migrateMacro:migrateMacro,
-		migrateRepeatingMacros:migrateRepeatingMacros,
-		migrateRepeatingMacrosMult: migrateRepeatingMacrosMult,
-		migrateStaticMacro: migrateStaticMacro,
-		migrateStaticMacros: migrateStaticMacros,
-		migrateStaticMacrosMult: migrateStaticMacrosMult
-	};
-}());
 var PFMenus = PFMenus || (function () {
 	'use strict';
 	/** creates a command macro button for a repeating section
@@ -1917,7 +344,7 @@ var PFMenus = PFMenus || (function () {
 		});
 	};
 	
-	console.log(PFLog.l + '   PFMenus module loaded          ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFMenus module loaded          ');
 	PFLog.modulecount++;
 	return {
 		resetOneCommandMacro: resetOneCommandMacro,
@@ -2136,7 +563,7 @@ var PFAbilityScores = PFAbilityScores || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFAbilityScores module loaded  ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFAbilityScores module loaded  ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -3354,7 +1781,7 @@ var PFMigrate = PFMigrate || (function () {
 		setAttrs(getAllMigrateFlags(), PFConst.silentParams, done);
 	}
 	;
-	console.log(PFLog.l + '   PFMigrate module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFMigrate module loaded        ');
 	PFLog.modulecount++;
 	return {
 		migrateRepeatingDamage: migrateRepeatingDamage,
@@ -3993,7 +2420,7 @@ var PFDefense = PFDefense || (function () {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFDefense module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFDefense module loaded        ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -4385,7 +2812,7 @@ var PFAttackGrid = PFAttackGrid || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFAttackGrid module loaded     ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFAttackGrid module loaded     ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -4573,7 +3000,7 @@ var PFAttackOptions = PFAttackOptions || (function () {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFAttackOptions module loaded  ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFAttackOptions module loaded  ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -5142,7 +3569,7 @@ var PFEncumbrance = PFEncumbrance || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFEncumbrance module loaded    ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFEncumbrance module loaded    ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -6618,7 +5045,7 @@ var PFInventory = PFInventory || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFInventory module loaded      ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFInventory module loaded      ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -7018,7 +5445,7 @@ var PFSpellOptions = PFSpellOptions || (function () {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFSpellOptions module loaded   ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFSpellOptions module loaded   ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -8555,7 +6982,7 @@ var PFSpells = PFSpells || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFSpells module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFSpells module loaded         ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -9171,7 +7598,7 @@ var PFSpellCasterClasses = PFSpellCasterClasses || (function () {
 		}); //end of spell classes
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + 'PFSpellCasterClasses module loaded' + PFLog.r, PFLog.bg);
+	PFConsole.log('PFSpellCasterClasses module loaded');
 	PFLog.modulecount++;
 	return {
 		ifSpellClassExists: ifSpellClassExists,
@@ -9503,7 +7930,7 @@ var PFClassRaceGrid = PFClassRaceGrid || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFClassRaceGrid module loaded  ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFClassRaceGrid module loaded  ');
 	PFLog.modulecount++;
 	return {
 		events: events,
@@ -10561,7 +8988,7 @@ var PFSkills = PFSkills || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFSkills module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFSkills module loaded         ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -11027,7 +9454,7 @@ var PFFeatures = PFFeatures || (function () {
 		
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFFeatures module loaded       ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFFeatures module loaded       ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -11995,7 +10422,7 @@ var PFAbility = PFAbility || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFAbility module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFAbility module loaded        ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -12791,7 +11218,7 @@ var PFAttacks = PFAttacks || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFAttacks module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFAttacks module loaded        ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -12887,7 +11314,7 @@ var PFPsionic = PFPsionic || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFPsionic module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFPsionic module loaded        ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -13011,7 +11438,7 @@ var PFMythic = PFMythic || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFMythic module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFMythic module loaded         ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -13337,7 +11764,7 @@ var PFHealth = PFHealth || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFHealth module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFHealth module loaded         ');
 	PFLog.modulecount++;
 	return {
 		migrate:migrate,
@@ -13455,7 +11882,7 @@ var PFSaves = PFSaves || (function () {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFSaves module loaded          ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFSaves module loaded          ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -13753,7 +12180,7 @@ var PFSize = PFSize || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFSize module loaded           ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFSize module loaded           ');
 	PFLog.modulecount++;
 	return {
 		migrate: migrate,
@@ -13801,7 +12228,7 @@ var PFInitiative = PFInitiative || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFInitiative module loaded     ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFInitiative module loaded     ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -13863,7 +12290,7 @@ var PFChecks = PFChecks || (function () {
 			}
 		});
 	};
-	console.log(PFLog.l + '   PFChecks module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFChecks module loaded         ');
 	PFLog.modulecount++;
 	return {
 		applyConditions: applyConditions
@@ -13978,7 +12405,7 @@ var PFConditions = PFConditions || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFConditions module loaded     ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFConditions module loaded     ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -14331,7 +12758,7 @@ var PFBuffs = PFBuffs || (function () {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFBuffs module loaded          ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFBuffs module loaded          ');
 	PFLog.modulecount++;
 	return {
 		recalculate: recalculate,
@@ -14386,7 +12813,7 @@ var PFNPC = PFNPC || (function () {
 		}));
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   PFNPC module loaded            ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   PFNPC module loaded            ');
 	PFLog.modulecount++;
 	return {
 		setToNPC: setToNPC,
@@ -17519,7 +15946,7 @@ var PFNPCParser = PFNPCParser || (function () {
 	registerEventHandlers = function () {
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   NPCParser module loaded        ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   NPCParser module loaded        ');
 	PFLog.modulecount++;
 	return {
 		importFromCompendium: importFromCompendium,
@@ -18739,7 +17166,7 @@ var HLImport = HLImport || (function() {
 		});
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '   HLImport module loaded         ' + PFLog.r, PFLog.bg);
+	PFConsole.log('   HLImport module loaded         ');
 	PFLog.modulecount++;
 	return {
 		importCharacter: importCharacter
@@ -19350,20 +17777,20 @@ var PFSheet = PFSheet || (function () {
 
 	};
 	registerEventHandlers();
-	console.log(PFLog.l + '       ,## /##                    ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '      /#/ /  ##                   ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '     / / /    ##                  ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '      | ##___#/                   ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '      | ##       athfinder        ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '   #  | ##    sheet version       ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '    ### /           ' + ("0000" + PFConst.version.toFixed(2)).slice(-5) + '         ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '                                  ' + PFLog.r, PFLog.bg);
-	console.log(PFLog.l + '   PFSheet module loaded          ' + PFLog.r, PFLog.bg);
+	PFConsole.log('       ,## /##                    ');
+	PFConsole.log('      /#/ /  ##                   ');
+	PFConsole.log('     / / /    ##                  ');
+	PFConsole.log('      | ##___#/                   ');
+	PFConsole.log('      | ##       athfinder        ');
+	PFConsole.log('   #  | ##    sheet version       ');
+	PFConsole.log('    ### /           ' + ("0000" + PFConst.version.toFixed(2)).slice(-5) + '         ');
+	PFConsole.log('                                  ');
+	PFConsole.log('   PFSheet module loaded          ');
 	PFLog.modulecount++;
 	if (PFLog.modulecount === 34) {
-		console.log(PFLog.l + '   All ' + PFLog.modulecount + ' Modules Loaded          ' + PFLog.r, PFLog.bg);
+		PFConsole.log('   All ' + PFLog.modulecount + ' Modules Loaded          ');
 	} else {
-		console.log(PFLog.l + '   ONLY ' + PFLog.modulecount + ' Modules Loaded!        ' + PFLog.r, 'background: linear-gradient(to right,yellow,white,white,yellow); color:black;text-shadow: 0 0 8px white;');
+		console.log(PFLog.l +'   ONLY ' + PFLog.modulecount + ' Modules Loaded!        ' + PFLog.r, 'background: linear-gradient(to right,yellow,white,white,yellow); color:black;text-shadow: 0 0 8px white;');
 	}
 	return {
 		recalculateCore: recalculateCore,
